+++ frontend/src/components/atoms/PltAmount.vue
@@ -1,7 +1,9 @@
 <template>
 	<span class="inline-block" data-testid="amount">
 		<span class="numerical">{{ amounts[0] }}</span>
-		<span v-if="amounts[1]" class="numerical text-sm opacity-50">.{{ amounts[1] }}</span>
+		<span v-if="amounts[1]" class="numerical text-sm opacity-50"
+			>.{{ amounts[1] }}</span
+		>
 	</span>
 </template>