+++ frontend/src/components/Charts/ChartBarST.vue
@@ -193,7 +193,7 @@ const handleBarClick = (event: MouseEvent) => {
 		if (y >= pixel - 10 && y <= pixel + 10 && x < yScale.left) {
 			const label = chartData.labels?.[i]
 			if (label) {
-				window.open(`/protocol-token/${label}`, '_blank')
+				window.open(`/protocol-token/${label}`, '_self')
 			}
 			return
 		}
@@ -210,7 +210,7 @@ const handleBarClick = (event: MouseEvent) => {
 		const index = elements[0].index
 		const label = chartData.labels?.[index] as string
 		if (label) {
-			window.open(`/protocol-token/${label}`, '_blank')
+			window.open(`/protocol-token/${label}`, '_self')
 		}
 	}
 }