+++ frontend/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Fixed
+
+- PLT overview page supply chart tooltip convert token symbol to token name.
+- Change the `Transaction` to `Activities` on the overview page of PLT and the per PLT page.
+
 ## [1.7.24] - 2025-08-29
 
 ### Fixed