+++ frontend/CHANGELOG.md
@@ -12,6 +12,10 @@ All notable changes to this project will be documented in this file.
 
 - Added token holder distribution chart to specific plt dashboards.
 
+### Fixed
+
+- fix validator `APY` section  `APY` figures were all overlapping resulting in them becoming unreadable.
+
 ## [1.7.24] - 2025-08-29
 
 ### Fixed