+++ frontend/CHANGELOG.md
@@ -16,6 +16,11 @@ All notable changes to this project will be documented in this file.
 
 - fix validator `APY` section  `APY` figures were all overlapping resulting in them becoming unreadable.
 
+### Fixed
+
+- PLT overview page supply chart tooltip convert token symbol to token name.
+- Change the `Transaction` to `Activities` on the overview page of PLT and the per PLT page.
+
 ## [1.7.24] - 2025-08-29
 
 ### Fixed