+++ frontend/CHANGELOG.md
@@ -2,7 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## Unreleased 
+
+### Fixed
+
+- Fixed issue with footer links not redirecting to proper telegram channel.
+
+### Added
+
+- Added token holder distribution chart to specific plt dashboards.
+
+### Fixed
+
+- fix validator `APY` section  `APY` figures were all overlapping resulting in them becoming unreadable.
+
+### Fixed
+
+- PLT overview page supply chart tooltip convert token symbol to token name.
+- Change the `Transaction` to `Activities` on the overview page of PLT and the per PLT page.
 
 ## [1.7.24] - 2025-08-29